Remove unused service includes
